refactor: split sync module into multiple files
